requirements.txt
packaging
ninja
numpy
scipy
torchao
diffusers
transformers
tokenizers
tqdm
accelerate
safetensors
opencv-python
imageio
imageio-ffmpeg
einops
loguru
sgl-kernel
qtorch
ftfy
gradio
aiohttp
pydantic
aio-pika
asyncpg>=0.27.0
aioboto3>=12.0.0
prometheus-client
gguf
fastapi
uvicorn
PyJWT
requests
alibabacloud_dypnsapi20170525==1.2.2
redis==6.4.0
tos
decord
zmq
jsonschema
pymongo
modelscope
